chore(home/chart): refactor TimeFilter
